+++ modules/interpreter/src/cpp/GeOperator.cpp
@@ -32,7 +32,7 @@ namespace Nelson {
 ArrayOf
 Evaluator::geOperator(AbstractSyntaxTreePtr t)
 {
-    callstack.pushID(t->getContext());
+    callstack.pushID((size_t)t->getContext());
     ArrayOf retval = this->geOperator(expression(t->down), expression(t->down->right));
     callstack.popID();
     return retval;
