+++ modules/interpreter/src/cpp/GtOperator.cpp
@@ -32,7 +32,7 @@ namespace Nelson {
 ArrayOf
 Evaluator::gtOperator(AbstractSyntaxTreePtr t)
 {
-    callstack.pushID(t->getContext());
+    callstack.pushID((size_t)t->getContext());
     ArrayOf retval = this->gtOperator(expression(t->down), expression(t->down->right));
     callstack.popID();
     return retval;
